+++ R/utilities-resolution.R
@@ -8,6 +8,8 @@
 #' @param x numeric vector
 #' @param zero should a zero value be automatically included in the
 #'   computation of resolution
+#' @param discrete should vectors mapped with a discrete scale be treated as
+#'   having a resolution of 1?
 #' @export
 #' @examples
 #' resolution(1:10)
@@ -17,9 +19,11 @@
 #' # Note the difference between numeric and integer vectors
 #' resolution(c(2, 10, 20, 50))
 #' resolution(c(2L, 10L, 20L, 50L))
-resolution <- function(x, zero = TRUE) {
-  if (is.integer(x) || is_mapped_discrete(x) ||
-      zero_range(range(x, na.rm = TRUE))) {
+resolution <- function(x, zero = TRUE, discrete = FALSE) {
+  if (is.integer(x) || zero_range(range(x, na.rm = TRUE))) {
+    return(1)
+  }
+  if (discrete && is_mapped_discrete(x)) {
     return(1)
   }